requirements.txt
Flask-SQLAlchemy==2.4.1
Flask==1.1.1
flask-JWT==0.3.2
flask-bcrypt==0.7.1
flask-marshmallow==0.10.1
marshmallow-sqlalchemy==0.19.0
psycopg2==2.8.4
pytest==5.3.1
testing.postgresql==1.3.0
